added components folder
